pyproject.toml
[tool.poetry]
name = "bsky-social-graph"
version = "0.1.0"
description = ""
authors = ["Victoriano Izquierdo <victoriano@gmail.com>"]
package-mode = false 


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
atproto = "^0.0.54"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
